requirements.txt
Flask==1.0.2
line-bot-sdk==1.15.0
beautifulsoup4==4.6.3
xlwt==1.3.0
lxml==4.2.5
